pyproject.toml
[project]
name = "pythonai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "faiss-cpu>=1.11.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "mcp[cli]>=1.10.1",
    "openai-whisper>=20250625",
    "pydantic>=2.11.7",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
    "watchdog>=6.0.0",
    "youtube-transcript-api>=1.1.1",
]
